Moved Exporter functions in models.py, fixed corresponding imports
